requirements.txt
python-telegram-bot==20.2
quorum-data-py
quorum-mininode-py  